requirements.txt
aiosmtplib==5.1.0
alembic==1.18.4
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
async-timeout==5.0.1
asyncpg==0.31.0
attrs==26.1.0
bcrypt==4.0.1
billiard==4.2.4
celery==5.6.3
celery-types==0.26.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.6
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.135.2
fastapi-cli==0.0.24
fastapi-cloud-cli==0.15.1
fastapi-mail==1.6.2
fastar==0.9.0
flower==2.0.1
graphql-core==3.2.8
greenlet==3.3.2
h11==0.16.0
harfile==0.4.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
humanize==4.15.0
hypothesis==6.152.1
hypothesis-graphql==0.12.0
hypothesis-jsonschema==0.23.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jsonschema_rs==0.46.0
junit-xml==1.9
kombu==5.6.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.1
passlib==1.7.4
pluggy==1.6.0
prometheus_client==0.25.0
prompt_toolkit==3.0.52
pyasn1==0.6.3
pycparser==3.0
pydantic==2.12.5
pydantic-extra-types==2.11.1
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pymongo==4.16.0
pyrate-limiter==4.1.0
pytest==9.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
regex==2025.11.3
requests==2.33.1
rich==14.3.3
rich-toolkit==0.19.7
rignore==0.7.6
rpds-py==0.30.0
rsa==4.9.1
schemathesis==4.15.2
sentry-sdk==2.56.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==0.52.1
starlette-testclient==0.4.1
tenacity==9.1.4
tornado==6.5.5
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.42.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
Werkzeug==3.1.8
# --- Core Framework (FastAPI aur Server) ---
fastapi[all]
uvicorn[standard]

# --- Database aur ORM ---
sqlalchemy
psycopg2-binary
alembic               # Database migrations ke liye (Is baar ye zaroori hai)

# --- Security aur Auth (Bookly wala logic) ---
python-jose[cryptography]
passlib[bcrypt]
python-multipart      # Form data aur files ke liye

# --- Configuration aur Env (Naya professional tarika) ---
pydantic-settings      # .env file ko smartly handle karne ke liye
python-dotenv

# --- Background Tasks aur Caching (15+ LPA level features) ---
redis
celery

# --- Testing ---
pytest
